Correctly handle tests that have not had recent runs
